# src/db-sql-courses/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Online Courses API",
    "description": "API для системи онлайн-курсів",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api"
    }
  ],
...
  "paths": {
...
    "/courses/high-rated/{minGrade}": {
      "get": {
        "summary": "Отримати курси з середнім балом > minGrade",
        "tags": ["Courses"],
        "parameters": [
          {
            "name": "minGrade",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Мінімальний бал"
          }
        ],
        "responses": {
          "200": {
            "description": "Список курсів"
          }
        }
      }
    }
  }
}